+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Added
 
+- Added Claude effort level and thinking settings
 - Properly handle results from agents running in background
 - Handle `Agent` tool type as an alias for `Task`
 - Handle display of TodoWrite tool